Moved all other clustering elements into scatter.ts
